.github/bmt/pyproject.toml
[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"

# Portable: copy this file and the `cli/` directory to any repo (e.g. prod); run `uv sync` in this directory.
# In CI, setup sets UV_PROJECT=.github/bmt so steps can use `uv run bmt <command>` without --project.
[tool.uv.build-backend]
module-name = "cli"
module-root = ""

[project]
name = "bmt"
version = "0.1.0"
requires-python = ">=3.12"
description = "BMT CI driver: matrix discovery, trigger, VM lifecycle, and runner upload."
dependencies = [
    "bmt-gcloud",
    "google-cloud-storage",
    "google-cloud-pubsub",
    "google-cloud-compute",
    "PyGithub",
    "pydantic",
]

[tool.uv.sources]
bmt-gcloud = { workspace = true }

[project.optional-dependencies]
dev = [
    "ruff",
    "basedpyright",
]

# Main CLI; shorthands below so you can run e.g. `uv run bmt-matrix` instead of `uv run bmt matrix`.
[project.scripts]
bmt = "cli.driver:main"
bmt-matrix = "cli.driver:main_matrix"
bmt-trigger = "cli.driver:main_write_run_trigger"
bmt-wait = "cli.driver:main_wait_handshake"
bmt-start-vm = "cli.driver:main_start_vm"
bmt-write-context = "cli.driver:main_write_context"
bmt-write-summary = "cli.driver:main_write_handoff_summary"
bmt-select-vm = "cli.driver:main_select_available_vm"

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [".venv", "__pycache__", ".ruff_cache"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.basedpyright]
typeCheckingMode = "basic"
pythonVersion = "3.12"
reportMissingTypeStubs = false
